+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,6 +19,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.GetAsync(request.Id);
+
+            if (leaveRequest is null)
+                throw new NotFoundException(nameof(LeaveRequest), request.Id);
+
             await _leaveRequestRepository.DeleteAsync(leaveRequest);
             return Unit.Value;
         }
